LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-constructor-expected.txt

Testing the MacOSFullscreenMediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "media-controls mac fullscreen"
PASS mediaControls.layoutTraits is an instance of MacOSLayoutTraits
PASS mediaControls.layoutTraits.isFullscreen is true

Controls bar
PASS mediaControls.bottomControlsBar.children.length is 4
PASS mediaControls.bottomControlsBar.children[1] is an instance of ButtonsContainer
PASS mediaControls.bottomControlsBar.children[2] is an instance of ButtonsContainer
PASS mediaControls.bottomControlsBar.children[3] is an instance of ButtonsContainer

Left container
PASS mediaControls.bottomControlsBar.children[1].children is [mediaControls.volumeSlider, mediaControls.muteButton]
PASS mediaControls.bottomControlsBar.children[1].element.className is "buttons-container left"
PASS mediaControls.bottomControlsBar.children[1].leftMargin is 12
PASS mediaControls.bottomControlsBar.children[1].rightMargin is 0
PASS mediaControls.bottomControlsBar.children[1].buttonMargin is 6

Center container
PASS mediaControls.bottomControlsBar.children[2].children is [mediaControls.rewindButton, mediaControls.playPauseButton, mediaControls.forwardButton]
PASS mediaControls.bottomControlsBar.children[2].element.className is "buttons-container center"
PASS mediaControls.bottomControlsBar.children[2].leftMargin is 27
PASS mediaControls.bottomControlsBar.children[2].rightMargin is 27
PASS mediaControls.bottomControlsBar.children[2].buttonMargin is 27

Right container
PASS mediaControls.bottomControlsBar.children[3].children is [mediaControls.airplayButton, mediaControls.pipButton, mediaControls.tracksButton, mediaControls.fullscreenButton, mediaControls.overflowButton]
PASS mediaControls.bottomControlsBar.children[3].element.className is "buttons-container right"
PASS mediaControls.bottomControlsBar.children[3].leftMargin is 12
PASS mediaControls.bottomControlsBar.children[3].rightMargin is 12
PASS mediaControls.bottomControlsBar.children[3].buttonMargin is 24

PASS successfullyParsed is true

TEST COMPLETE
